✨ (ghub.json): update package versions for crush, opencode, lazygit, zed, LocalAI, uv, and bun 🔧 (userprefs.conf): update lockscreen script path and comment out monitors.conf source 🔧 (mpv.conf): remove watch-later-options-min-playtime setting ✨ (mod.nu): add hyde-monitor command and update hyde command to not include monitors.conf by default ♻️ (ctx.nu): refactor path-top function and comment out add file functionality ✨ (ghub/mod.nu): add condition to check if changelog should be shown 🔧 (install-base.nu): update environment variables and add CONFIG_DIRS 🔧 (install-desktop.nu): update environment variables and add CONFIG_DIRS ✨ (pacman.nu): add zip and resvg to tools and yazi dependencies ♻️ (preview.nu): refactor fuzzy function to handle execute commands better ✨ (tdown.nu): add max-size and max-dur options to download command ⬆️ (nushell.sh): update nushell version to 0.106.1 📝 (taskell.md): update task list with new items

+++ modules/ghub/mod.nu
@@ -429,7 +429,7 @@ export def "repo update" [...names: string@names, --changelog(-c), --loop(-l), -
     $current_index = $current_index + 1
   }
 
-  if (ls $changelog_dir | is-not-empty) {
+  if $save_changelog and (ls $changelog_dir | is-not-empty) {
     if (confirm "see changelog...?") {
       glow $changelog_dir
       rm -rfp $changelog_dir
